pyproject.toml
[project]
name = "anthropic-cookbook"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "llama-index>=0.14.20",
    "llama-index-embeddings-huggingface>=0.7.0",
    "llama-index-llms-anthropic>=0.11.2",
    "python-dotenv>=1.0.0",
    "sentence-transformers>=3.0.0",
    "numpy>=1.26.0",
]
